refactor and cleanup based on the review comments
