+++ config/300-buildtemplate.yaml
@@ -29,6 +29,3 @@ spec:
   - name: Age
     type: date
     JSONPath: .metadata.creationTimestamp
-  - name: Completed
-    type: date
-    JSONPath: .status.completionTime
